Tnext() was completely wrong.
It isn't possible (or at least not straightforward) to find the
lexicographically next key from the trie given a string that is
not present in the trie. We don't know when scanning down the
trie at which point our string differs from the keys in this
subtrie - it could be at some untested point between branch
nodes, and our string could be less or greater than the keys
in the subtrie. So our search can go off the rails in arbitrary
directions.

But, given a key that is present in the trie, we can search
forwards for the immediate successor.