Public suffix list compression experiment
The idea is to use an application-specific trie to store the PSL, reducing
the alphabet to LDH plus '.' and '\0', using bitmap compression of nodes,
not omitting intermediate branchless nodes, and storing the trailing non-
branchy part of a key as a string rather than a chain of nodes. PSL strings
must be converted from UTF-8 to punycode and reversed before looking them up.

Branch nodes contain a 1 bit, a 39 bit map, and a 16 bit offset to the twigs
(7 bytes). Leaf strings start with a 0 bit and are '\0' terminated ASCII.

psl_count.pl suggests we might be able to store the PSL in
about 3.5 Kbytes.