Re-jig the old twig deletion code to be more rib-friendly.
The old logic was based on the twig array index of the twig
to be deleted (called 's' in the code).

The rib compression code needs to delete a twig from the middle
of a heterogeneous allocation, which is easier to do knowing
the address and size of the allocation and the address and size
of the part to be deleted.

This lower-level view actually works pretty well in the old code,
since 't' (the pointer to the twig to be deleted) is equivalent to
's', so we can use it directly instead of re-calculating the
array index with another popcount.