DNS: an iterative Tnext()
Instead of recursively traversing the tree, walk down it while keeping
track of our target's immediate neighbours. The tree walk is very
similar to Tset(): we have to do a preliminary walk to find a
leaf (which will go off piste if our search key is not in the tree),
work out out the common prefix of the keys, then do a second walk with
the correct termination conditions.